Clear the current time of dynamical systems before re-simulation. Fixes #46 [skip ci]
